refactor(#2453 R2): split App-arm handler tests into restart_tests.rs (LOC gate)
The R2 App-arm handler tests pushed src/mcp/handlers/restart.rs to 826 LOC, over
the file_size_invariant MAX_LOC (750) that guards against MCP-handler monolith
regrowth. Move the two test modules verbatim into a sibling restart_tests.rs (its
name contains "test" → the invariant skips it), included via
`#[cfg(test)] #[path = "restart_tests.rs"] mod restart_tests;`. The split-out
modules become grandchildren of `restart`, so their imports use `super::super::*`.
restart.rs → 709 LOC. Tests unchanged (now at restart::restart_tests::…) and pass;
mcp_handler_files_under_max_loc + the merge-freshness pin pass. clippy/fmt clean.


Agend-Agent: archfix-claude-opus
Agend-Task: t-20260712034222302055-19288-5
Agend-Branch: feat/app-staged-restart-r2
Agend-Issued-At: 2026-07-12T03:43:18.424715+00:00
